requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
Flask-JWT-Extended==4.7.1
Flask-Mailman==1.0.0
Flask-Migrate==4.0.7
marshmallow==3.23.1
flask-marshmallow==1.2.1
bcrypt==4.2.0
cryptography==44.0.0
mercadopago==2.3.0
google-genai==1.0.0
APScheduler==3.10.4
Pillow==11.0.0
reportlab==4.2.5
flask-smorest==0.45.0
Flask-Limiter==3.9.0
Flask-Cors==4.0.1
Flask-Talisman==1.1.0
python-dotenv==1.0.1
supabase==2.10.0
cloudinary==1.41.0
pytest==8.3.4
pytest-flask==0.15.1
psycopg2-binary==2.9.10
gunicorn==23.0.0
